users_function/function_get_users/requirements.txt
sqlalchemy
sqlalchemy-utils
psycopg2-binary
azure-functions
